refactor(example): split system.go transform into per-concern files
